requirements.txt
asgiref==3.11.0
beautifulsoup4==4.14.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework-gis==1.2.0
docopt==0.6.2
idna==3.11
Js2Py==0.74
numpy==2.3.5
packaging==25.0
pipwin==0.5.2
psycopg2-binary==2.9.11
pyjsparser==2.7.1
PyPrind==2.11.3
pySmartDL==1.3.4
requests==2.32.5
scipy==1.16.3
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
